rename : sample class from test
