+++ build.gradle.kts
@@ -81,7 +81,6 @@ project(":service") {
 
 project(":entity") {
     dependencies {
-        // 실제 DB 연결할 때 디펜던시 추가할 것
         implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     }
 }
