+++ entity/src/main/kotlin/com/mashup/dojo/SampleEntity.kt
@@ -1,0 +1,4 @@
+package com.mashup.dojo
+
+class SampleEntity {
+}
