--- service/src/main/kotlin/com/mashup/dojo/domain/DojoDomain.kt
@@ -1,10 +1,0 @@
-package com.mashup.dojo.domain
-
-class DojoDomain(
-    private val name: String
-) {
-
-    fun name(): String {
-        return name
-    }
-}
