+++ service/src/main/kotlin/com/mashup/dojo/domain/Sample.kt
@@ -1,0 +1,6 @@
+package com.mashup.dojo.domain
+
+data class Sample(
+    val id: Long,
+    val name: String
+)
