+++ service/src/main/kotlin/com/mashup/dojo/service/SampleService.kt
@@ -1,0 +1,12 @@
+package com.mashup.dojo.service
+
+import com.mashup.dojo.domain.Sample
+import org.springframework.stereotype.Service
+
+@Service
+class SampleService {
+    fun getSampleId(): Long {
+        val sample = Sample(1, "do~jo")
+        return sample.id
+    }
+}
