+++ service/src/main/kotlin/com/mashup/dojo/usecase/SampleUseCase.kt
@@ -1,0 +1,13 @@
+package com.mashup.dojo.usecase
+
+import com.mashup.dojo.service.SampleService
+import org.springframework.stereotype.Component
+
+@Component
+class SampleUseCase(
+    private val sampleService: SampleService
+) {
+    fun getSampleId(): Long {
+        return sampleService.getSampleId()
+    }
+}
